ALL | SUSPENDING PROJECT FOR NOW
- have to suspend my work on this competition in favor of a more important project which is due on 19th of November, a react app.
- currently only bud.cpp is correctly written, and zap.cpp is closest to being completed as well, I'm sure I will manage to perhaps do some more work on it before the deadline, but I am not certain i will be able to finish it completely.

+++ cza.cpp
@@ -24,12 +24,14 @@
 // S - the n-long string, (a-z)*n
 // CLARIFICATIONS:
 // as far as i understand, S is already polluted on input with R of length k placed on the end,
-// what i don't understand is why is S' introduced then right at the very beggining of the example
+// what i don't understand is why is S' size_troduced then right at the very beggining of the example
 // if it's just equal to regular S, S = S'
 // * We isolate the R by taking last k characters out of the raw S.
 // * We index every previous occurence of R, as well as the character which appeared straight after it.
 // * We take the most common character appearing right after our R's occurences, we append that letter to the end.
 // * We reisolate the R from our new S, then repeat this process until we extend S enough to output it's values specified by a and b variables.
+
+// TODO: UPDATE 3/11/2023: WE SHOULD INDEX EVERY PATTERN WHICH IS QUERIED, ALSO ADD ANY NEW PATTERNS TO THAT INDEX, THIS WILL MAKE THIS SO FAST 
 
 int main() {
 	unsigned long long stringLength, suffixLength, outputIndexStart, outputIndexEnd;
@@ -41,14 +43,14 @@ int main() {
 
 	// convert input to vector form, todo: test performance of this vs direct std::string	
 	std::vector<char> baseString(stringLength);
-	for (int i = 0; i < stringLength; i++) {
+	for (size_t i = 0; i < stringLength; i++) {
 		baseString[i] = inputString.at(i);	
 	}
 
 	do {
 		// find current suffix | verified
 		std::vector<char> currentSuffix(suffixLength);
-		for (int i = baseString.size() - suffixLength, j = 0; j < suffixLength; i++, j++) {
+		for (size_t  i = baseString.size() - suffixLength, j = 0; j < suffixLength; i++, j++) {
 			currentSuffix[j] = baseString[i];
 			//std::cout << currentSuffix[j];
 		}
@@ -56,11 +58,11 @@ int main() {
 
 		// find each occurence of suffix | verified 
 		// (the hardcode is ugly but it will be easier to test and debug it later)
-		std::map<char, int> occurenceCounter {{'a', 0}, {'b', 0}, {'c', 0}, {'d', 0}, {'e', 0}, {'f', 0}, {'g', 0}, {'h', 0}, {'i', 0}, {'j', 0}, {'k', 0}, {'l', 0}, {'m', 0}, {'n', 0}, {'o', 0}, {'p', 0}, {'q', 0}, {'r', 0}, {'s', 0}, {'t', 0}, {'u', 0}, {'v', 0}, {'w', 0}, {'x', 0}, {'y', 0}, {'z', 0}};
-		for (int i = 0; i < baseString.size(); i++) {
+		std::map<char, size_t> occurenceCounter {{'a', 0}, {'b', 0}, {'c', 0}, {'d', 0}, {'e', 0}, {'f', 0}, {'g', 0}, {'h', 0}, {'i', 0}, {'j', 0}, {'k', 0}, {'l', 0}, {'m', 0}, {'n', 0}, {'o', 0}, {'p', 0}, {'q', 0}, {'r', 0}, {'s', 0}, {'t', 0}, {'u', 0}, {'v', 0}, {'w', 0}, {'x', 0}, {'y', 0}, {'z', 0}};
+		for (size_t i = 0; i < baseString.size(); i++) {
 			//std::cout << "checking occurence for i = " << i << " : ";
 			bool isPresent = true;
-			for (int suffixIdx = 0; suffixIdx < suffixLength && i + suffixIdx < baseString.size(); suffixIdx++) {
+			for (size_t suffixIdx = 0; suffixIdx < suffixLength && i + suffixIdx < baseString.size(); suffixIdx++) {
 				//std::cout << baseString[suffixIdx + i];
 				if (baseString[suffixIdx + i] != currentSuffix[suffixIdx]) {
 					//std::cout << " terminated: " << baseString[suffixIdx + i] << "!=" << currentSuffix[suffixIdx];
@@ -82,7 +84,7 @@ int main() {
 		
 		// find most common letter of all | verified
 		char mostCommonLetter = 'a'; // apparently if no other option is available, 'a' takes the priority as it's the first letter of the set of all letters equally eligible
-		int mostOccurenceCount = 0;
+		size_t mostOccurenceCount = 0;
 		for (auto it = occurenceCounter.begin(); it != occurenceCounter.end(); it++) {
 			//std::cout << it->first << ':' << it->second << std::endl;
 			// note: earlier letter in the alphabet takes priority in case of ties
@@ -107,7 +109,7 @@ int main() {
 	// ^^^ code above does not change runtime speed too much, will test it with the last test if i manage to get it working
 
 	std::string outputString;
-	for (int i = outputIndexStart - 1; i < outputIndexEnd; i++) {
+	for (size_t i = outputIndexStart - 1; i < outputIndexEnd; i++) {
 		outputString.push_back(baseString[i]);
 		// std::cout << i << ':' << baseString[i] << std::endl;
 	}
